version 2.3.4

+++ backend/routes/cart.py
@@ -1,9 +1,7 @@
-import os
-from typing import Optional
+from datetime import timedelta
 from datetime import datetime 
 
-from fastapi_pagination import Page, paginate
-from fastapi import APIRouter, Depends, HTTPException, UploadFile, File
+from fastapi import APIRouter, Depends, HTTPException, BackgroundTasks
 from services.dbconnection import GetDB
 from sqlalchemy.orm import session
 from services.userscrud import GetUserID
@@ -19,6 +17,7 @@ from schemas.cart import CarItemCreate, CartItemDataResponse, CartItemBase, Cart
 from sqlalchemy.orm import joinedload
 
 cart_routes = APIRouter()
+background_tasks = BackgroundTasks();
 
 """obtener los productos del carrito"""
 @cart_routes.get("/cart/{current_user}", response_model=CartDataResponse)
@@ -42,51 +41,75 @@ def get_cart(current_user : int, db : session = Depends(GetDB), token : str=Depe
     
     return cart    
 
+"""reserva por 48horas"""
+def release_product_reservation(product_id: int, quantity: int, db: session):
+    product = db.query(Product).filter(Product.id == product_id).first()
+    if product:
+        product.reserved_quantity -= quantity
+        db.add(product)
+        db.commit()
+
 """agregar productos al carrito"""
 @cart_routes.post("/cart/items/{current_user}", response_model=CartItemDataResponse)
-def add_item_cart(current_user : int ,item_data : CarItemCreate,db : session = Depends(GetDB), token : str=Depends(oauth2_scheme)):
+def add_item_cart(current_user: int, item_data: CarItemCreate, db: session = Depends(GetDB), token: str = Depends(oauth2_scheme)):
     decoded_token = decode_token(token)
     user = GetUserID(db, decoded_token["id"])
-    if user.role not in ["admin", "client"]:
-        raise HTTPException(status_code=403, detail="Not authorized to add items cart")
     
-    query = db.query(Cart).filter(Cart.user_id == current_user).first()
-    if not query:
+    if user.role not in ["admin", "client"]:
+        raise HTTPException(status_code=403, detail="Not authorized to add items to cart")
+    
+    # Recuperar el carrito del usuario
+    cart = db.query(Cart).filter(Cart.user_id == current_user).first()
+    
+    if not cart:
         cart = Cart(user_id=current_user, created_at=datetime.utcnow(), total_value=0.0)
         db.add(cart)
         db.commit()
         db.refresh(cart)
-    else:
-        cart = query
-    # Check if the product exists
+
+    # Verificar si el producto existe
     product = db.query(Product).filter(Product.id == item_data.product_id).first()
+    
     if not product:
         raise HTTPException(status_code=404, detail="Product not found")
     
-    # Retrieve or create the cart item
+    # Verificar si hay suficiente cantidad disponible
+    available_quantity = product.quantity - product.reserved_quantity
+    if item_data.quantity > available_quantity:
+        raise HTTPException(status_code=400, detail="Not enough stock available")
+
+    # Actualizar el campo quantity del producto
+    product.quantity -= item_data.quantity  # Resta del total disponible
+    product.reserved_quantity += item_data.quantity  # Incrementa la cantidad reservada
+    db.add(product)  # Guarda los cambios del producto
+    
+    # Recuperar o crear el ítem del carrito
     existing_item = db.query(CartItem).filter(
         CartItem.cart_id == cart.id,
         CartItem.product_id == item_data.product_id
     ).first()
     
     if existing_item:
-        existing_item.quantity = item_data.quantity
-        existing_item.total_price = existing_item.quantity * product.price_product  # Calculate total price
+        existing_item.quantity += item_data.quantity  # Suma a la cantidad existente
+        existing_item.total_price = existing_item.quantity * product.price_product  # Actualiza el precio total
         new_item = existing_item
     else:
         new_item = CartItem(
             cart_id=cart.id,
             product_id=item_data.product_id,
             quantity=item_data.quantity,
-            total_price=item_data.quantity * product.price_product  # Calculate total price based on quantity
+            total_price=item_data.quantity * product.price_product  # Precio total según cantidad
         )
         db.add(new_item)
-    
-    # Update the total value of the cart
+
+    # Actualizar el total del carrito
     cart.total_value = db.query(func.sum(CartItem.total_price)).filter(CartItem.cart_id == cart.id).scalar() or 0.0
-    
+
     try:
-        db.commit()
+        # Agregar tarea en segundo plano para liberar la reserva después de 48 horas
+        background_tasks.add_task(release_product_reservation, product.id, item_data.quantity, db)
+
+        db.commit()  # Realiza el commit aquí después de todos los cambios
         db.refresh(cart)
         return new_item
     except IntegrityError as e:
@@ -96,46 +119,91 @@ def add_item_cart(current_user : int ,item_data : CarItemCreate,db : session = D
 
 """actualizar producto en la lista del carrito de compra"""
 @cart_routes.put("/cart/items/{item_id}", response_model=CartItemDataResponse)
-def update_item_quantity(item_id : int, item_data : CartItemBase, current_user : int, db: session = Depends(GetDB), token : str = Depends(oauth2_scheme)):
+def update_item_quantity(item_id: int, item_data: CartItemBase, current_user: int, db: session = Depends(GetDB), token: str = Depends(oauth2_scheme)):
     decoded_token = decode_token(token)
     user = GetUserID(db, decoded_token["id"])
     if user.role not in ["admin", "client"]:
-        raise HTTPException(status_code=403, detail="Not authorized to add items cart")
+        raise HTTPException(status_code=403, detail="Not authorized to update items in cart")
     
     query = db.query(CartItem).join(Cart).filter(CartItem.id == item_id, Cart.user_id == current_user)
     item = query.first()
-    
+
     if not item:
-        raise HTTPException(status_code=404, detail="Not item found in the cart")
+        raise HTTPException(status_code=404, detail="Item not found in the cart")
+
+    product = db.query(Product).filter(Product.id == item.product_id).first()
     
+    if not product:
+        raise HTTPException(status_code=404, detail="Product not found")
+
+    # Calcular la diferencia en la cantidad solicitada
+    quantity_diff = item_data.quantity - item.quantity
+
+    # Verificar si hay suficiente stock si la cantidad está aumentando
+    if quantity_diff > 0:
+        available_stock = product.quantity - product.reserved_quantity
+        if quantity_diff > available_stock:
+            raise HTTPException(status_code=400, detail="Not enough stock available")
+        # Aumentar la cantidad reservada en el stock
+        product.reserved_quantity += quantity_diff
+    else:
+        # Devolver la diferencia al stock
+        product.reserved_quantity += quantity_diff  # Aquí quantity_diff es negativo
+
     item.quantity = item_data.quantity
-    item.total_price = item_data.quantity * Product.price_product
+    item.total_price = item_data.quantity * product.price_product
     
+    # Actualizar el valor total del carrito
     cart = db.query(Cart).filter(Cart.id == item.cart_id).first()
     cart.total_value = db.query(func.sum(CartItem.total_price)).filter(CartItem.cart_id == cart.id).scalar()
-    
+
     db.commit()
     db.refresh(item)
     return item
 
 """eliminar un elemento del carrito de compra"""
 @cart_routes.delete("/cart/items/delete/{item_id}", response_model=dict)
-def delete_item_from_cart(item_id : int, current_user : int, db : session = Depends(GetDB), token : str = Depends(oauth2_scheme)):
+def delete_item_from_cart(item_id: int, current_user: int, db: session = Depends(GetDB), token: str = Depends(oauth2_scheme)):
     decoded_token = decode_token(token)
     user = GetUserID(db, decoded_token["id"])
-    if user.role not in ["admin", "client"]:
-        
-        raise HTTPException(status_code=403, detail="Not authorized to add items cart")
     
+    if user.role not in ["admin", "client"]:
+        raise HTTPException(status_code=403, detail="Not authorized to delete items from cart")
+    
+    # Obtener el item del carrito del usuario actual
     item = db.query(CartItem).join(Cart).filter(CartItem.id == item_id, Cart.user_id == current_user).first()
 
     if not item:
         raise HTTPException(status_code=404, detail="Item not found in the cart")
+
+    # Recuperar el producto correspondiente
+    product = db.query(Product).filter(Product.id == item.product_id).first()
+
+    if not product:
+        raise HTTPException(status_code=404, detail="Product not found")
+
+    # Restar la cantidad reservada por el usuario al reserved_quantity del producto
+    product.reserved_quantity -= item.quantity  # Aquí se resta la cantidad del item que se está eliminando
+
+    # Verificar que reserved_quantity no sea menor que 0
+    if product.reserved_quantity < 0:
+        product.reserved_quantity = 0  # Asegúrate de que no se vuelva negativo
+
+
+    # Sumar la cantidad del item eliminado al quantity del producto
+    product.quantity += item.quantity  # Aquí se suma la cantidad que se está eliminando
+
+    # Actualizar el producto en la base de datos
+    db.add(product)
+
+    # Eliminar el item del carrito
     db.delete(item)
 
-    # Update the cart's total value
+    # Actualizar el valor total del carrito
     cart = db.query(Cart).filter(Cart.id == item.cart_id).first()
     cart.total_value = db.query(func.sum(CartItem.total_price)).filter(CartItem.cart_id == cart.id).scalar()
 
-    db.commit()
+    db.commit()  # Asegúrate de hacer commit para guardar todos los cambios
     return {"detail": "Item deleted successfully"}
+
+
